separate testapp codebase
